feat: balance endpoint, change quantity to decimal

+++ trading-platform/Dtos/OrderDto.cs
@@ -4,7 +4,7 @@
     {
         public int UserId { get; set; }
         public int StockId { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string Type { get; set; } = string.Empty;
     }
     public class OrderResponseDto
@@ -12,7 +12,7 @@
         public int Id { get; set; }
         public int UserId { get; set; }
         public int StockId { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string Type { get; set; } = string.Empty;
         public DateTime Timestamp { get; set; }
         public decimal PriceAtExecution { get; set; }
